test : 추가 작성한 테스트 메서드 SelfMadeTest 클래스로 이전
